Generate enums in header file instead
